fix building 32 bits builds - rapidjson does not know what to make of an unsigned long
git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns@2968 d19b8d6e-7fed-0310-83ef-9ca221ded41b

+++ pdns/ws.cc
@@ -502,7 +502,7 @@ string StatWebServer::jsonstat(const string& method, const string& post, const m
       jdi.AddMember("masters", jmasters, doc.GetAllocator()); // ^^^ this makes an actual copy, otherwise the zerocopy behaviour bites us!
       jdi.AddMember("serial", di.serial, doc.GetAllocator());
       jdi.AddMember("notified_serial", di.notified_serial, doc.GetAllocator());
-      jdi.AddMember("last_check", di.last_check, doc.GetAllocator());
+      jdi.AddMember("last_check", (unsigned int) di.last_check, doc.GetAllocator());
       jdomains.PushBack(jdi, doc.GetAllocator());
     }
     doc.AddMember("domains", jdomains, doc.GetAllocator());
